requirements.txt
streamlit
azure-ai-contentsafety